backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
alembic==1.13.1
gunicorn==22.0.0

# LLM and RAG
llama-index==0.10.46
llama-index-vector-stores-chroma==0.1.8
llama-index-llms-ollama==0.1.4
llama-index-embeddings-ollama==0.1.2
chromadb-client==0.5.0
# llama-index-embeddings-huggingface==0.2.0 